fianl cmaog

+++ FRDM-TFC/Project_Headers/TFC/TFC_ADC.h
@@ -10,7 +10,7 @@ void ADC0_IRQ();
 
 float TFC_ReadPot(uint8_t Channel);
 float TFC_ReadBatteryVoltage();
-uint16_t TFC_ReadHBridgeFeedBack(uint8_t Channel);
+float TFC_ReadHBridgeFeedBack(uint8_t Channel);
 
 #endif /* __ADC16_H__ */
